chore: add nav-links component
