+++ package.json
@@ -18,7 +18,6 @@
     "next-auth": "^5.0.0-beta.19",
     "react": "^18",
     "react-dom": "^18",
-    "swr": "^2.2.5",
     "zod": "^3.23.8"
   },
   "devDependencies": {
