+++ src/app/ui/partials/index.tsx
@@ -1,3 +1,4 @@
 import Sidebar from './sidebar';
+import NavLinks from './nav-links';
 
-export { Sidebar };
+export { Sidebar, NavLinks };
