+++ src/types/menu.ts
@@ -4,9 +4,8 @@ export interface Menu {
   id?: number;
   name: string;
   href: string;
-  current: boolean;
   icon?: any;
   initial?: string;
-  children?: Menu[];
   count?: number;
+  children?: Menu[];
 }
